improve call to struct.calcsize
struct.calcsize calls malloc on fmt. If count is large because of a corrupt file, this may not have been caught yet, and could result in a large chunk of memory being allocated unnecessarily. Moving count outside of the call to struct.calcsize still gives the correct size, but can be a significantly smaller malloc.